services/movies_admin/movies_admin/movies/api/v1/urls.py
from django.urls import path

from .views import MoviesListApi, MoviesDetailApi

urlpatterns = [
    path('movies/', MoviesListApi.as_view()),
    path('movies/<uuid:pk>', MoviesDetailApi.as_view()),
]
