services/movies_admin/movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
django-extensions==3.1.0